+++ src/linear/models/__init__.py
@@ -4,7 +4,15 @@ from .base import Organization, PageInfo
 from .comments import Comment as CommentModel
 from .comments import CommentConnection, CommentUser
 from .cycles import Cycle, CycleConnection
-from .issues import Attachment, Comment, Issue, IssueConnection, WorkflowState
+from .issues import (
+    Attachment,
+    Comment,
+    Issue,
+    IssueConnection,
+    IssueRelation,
+    IssueRelationConnection,
+    WorkflowState,
+)
 from .labels import Label, LabelConnection
 from .projects import Project, ProjectConnection
 from .teams import Team, TeamConnection
@@ -24,6 +32,8 @@ __all__ = [
     "Attachment",
     "Issue",
     "IssueConnection",
+    "IssueRelation",
+    "IssueRelationConnection",
     # Projects
     "Project",
     "ProjectConnection",